Allow packages as originating elements
If a type is created from a package declaration
then that type should react to changes in the
package-info.java file. There is no need to forbid
packages as originating elements.

This also extracts the top level element logic
into a helper class, as it will soon be needed
in other places too for multiple-origin annotation
processors.